# contracts/http/v1/openapi.yaml
openapi: 3.1.0
info:
  title: AegisLink HTTP API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:4321
security:
  - sessionCookie: []
paths:
  /healthz:
    get:
      operationId: health
      security: []
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /readyz:
    get:
      operationId: readiness
      security: []
      responses:
        "200":
          description: Dependencies are ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "503":
          $ref: "#/components/responses/Error"
  /api/v1/auth/register:
    post:
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account, Human Principal, Personal Agent, and session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated account and new server-side session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/logout:
    post:
      operationId: logout
      security: []
      responses:
        "204":
          description: Session revoked and cookie cleared
  /api/v1/auth/session:
    get:
      operationId: getSession
      responses:
        "200":
          description: Current authenticated User and Personal Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/bootstrap:
    get:
      operationId: bootstrap
      responses:
        "200":
          description: Authenticated User, Personal Agent, and model metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapResponse"
  /api/v1/ag-ui:
    post:
      operationId: runAgent
      summary: Run the personal agent over the AG-UI protocol
      description: |
        Accepts an AG-UI RunAgentInput and streams AG-UI events. The threadId
        must identify an existing AegisLink conversation. PostgreSQL history is
        authoritative; earlier client messages are not used as trusted history.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgUiRunAgentInput"
      responses:
        "200":
          description: AG-UI Server-Sent Event stream
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/agents:
    get:
      operationId: listAgents
      responses:
        "200":
          description: Personal Agents owned by the authenticated Principal
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Agent"
  /api/v1/agents/{agentId}/memories:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    get:
      operationId: listAgentMemories
      responses:
        "200":
          description: Active long-term memories scoped to this Personal Agent
          content:
            application/json:
              schema:
                type: object
                required: [memories]
                properties:
                  memories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Memory"
        "404":
          $ref: "#/components/responses/Error"
    post:
      operationId: createAgentMemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryRequest"
      responses:
        "201":
          description: Agent-scoped memory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/agents/{agentId}/memories/{memoryId}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
      - $ref: "#/components/parameters/MemoryId"
    patch:
      operationId: updateAgentMemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemoryRequest"
      responses:
        "200":
          description: Memory updated or confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      operationId: forgetAgentMemory
      responses:
        "204":
          description: Memory forgotten and excluded from future retrieval
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/agents/{agentId}/skills:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    get:
      operationId: listAgentSkills
      responses:
        "200":
          description: Skills installed for this Personal Agent
          content:
            application/json:
              schema:
                type: object
                required: [skills]
                properties:
                  skills:
                    type: array
                    items:
                      $ref: "#/components/schemas/Skill"
        "404":
          $ref: "#/components/responses/Error"
    post:
      operationId: installAgentSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallSkillRequest"
      responses:
        "201":
          description: SKILL.md validated, installed, and bound to this Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/agents/{agentId}/skills/{skillId}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
      - $ref: "#/components/parameters/SkillId"
    patch:
      operationId: updateAgentSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: Per-Agent Skill enablement updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      operationId: uninstallAgentSkill
      responses:
        "204":
          description: Skill unbound from this Agent
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/agents/{agentId}/mcp-servers:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    get:
      operationId: listAgentMcpServers
      responses:
        "200":
          description: MCP Servers configured for this Personal Agent
          content:
            application/json:
              schema:
                type: object
                required: [servers]
                properties:
                  servers:
                    type: array
                    items:
                      $ref: "#/components/schemas/McpServer"
        "404":
          $ref: "#/components/responses/Error"
    post:
      operationId: createAgentMcpServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMcpServerRequest"
      responses:
        "201":
          description: Agent-scoped MCP Server configuration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServer"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/agents/{agentId}/mcp-servers/{serverId}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
      - $ref: "#/components/parameters/McpServerId"
    patch:
      operationId: updateAgentMcpServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: MCP Server enablement updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServer"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteAgentMcpServer
      responses:
        "204":
          description: MCP Server and cached tool catalog removed from this Agent
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/agents/{agentId}/mcp-servers/{serverId}/refresh:
    parameters:
      - $ref: "#/components/parameters/AgentId"
      - $ref: "#/components/parameters/McpServerId"
    post:
      operationId: refreshAgentMcpServer
      responses:
        "200":
          description: MCP connection checked and tool catalog refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServer"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
  /api/v1/agents/{agentId}/mcp-servers/{serverId}/tools/{toolName}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
      - $ref: "#/components/parameters/McpServerId"
      - name: toolName
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateAgentMcpTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled, riskLevel]
              properties:
                enabled:
                  type: boolean
                riskLevel:
                  type: string
                  enum: [read-only, external-write, destructive]
      responses:
        "200":
          description: Per-Agent MCP Tool permission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServer"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/conversations:
    get:
      operationId: listConversations
      responses:
        "200":
          description: Conversations ordered by latest activity
          content:
            application/json:
              schema:
                type: object
                required: [conversations]
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 120
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                type: object
                required: [conversation]
                properties:
                  conversation:
                    $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/conversations/{conversationId}:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
    get:
      operationId: getConversation
      responses:
        "200":
          description: Conversation and ordered messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/conversations/{conversationId}/messages:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
    post:
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 32768
      responses:
        "202":
          description: Message persisted and run queued
          content:
            application/json:
              schema:
                type: object
                required: [message, run]
                properties:
                  message:
                    $ref: "#/components/schemas/Message"
                  run:
                    $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/runs/{runId}/events:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: streamRunEvents
      parameters:
        - name: Last-Event-ID
          in: header
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Server-sent stream of RunEvent objects
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/runs/{runId}/cancel:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      operationId: cancelRun
      responses:
        "202":
          description: Cancellation requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: aegislink_session
  parameters:
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
    MemoryId:
      name: memoryId
      in: path
      required: true
      schema:
        type: string
    SkillId:
      name: skillId
      in: path
      required: true
      schema:
        type: string
    McpServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
    RegisterRequest:
      type: object
      required: [displayName, email, password]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 10
          maxLength: 128
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 128
    User:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
        displayName:
          type: string
    AuthResponse:
      type: object
      required: [user, agent]
      properties:
        user:
          $ref: "#/components/schemas/User"
        agent:
          $ref: "#/components/schemas/Agent"
    Agent:
      type: object
      required: [id, name, description, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Memory:
      type: object
      required: [id, agentId, kind, content, confidence, sourceUri, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        agentId:
          type: string
        kind:
          type: string
          enum: [semantic, episodic]
        content:
          type: string
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        sourceUri:
          type: string
        status:
          type: string
          enum: [active, forgotten]
        lastConfirmedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateMemoryRequest:
      type: object
      required: [kind, content, sourceUri]
      properties:
        kind:
          type: string
          enum: [semantic, episodic]
        content:
          type: string
          minLength: 1
          maxLength: 4000
        sourceUri:
          type: string
          maxLength: 1000
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          default: 1
    UpdateMemoryRequest:
      type: object
      properties:
        kind:
          type: string
          enum: [semantic, episodic]
        content:
          type: string
          minLength: 1
          maxLength: 4000
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        confirmed:
          type: boolean
    Skill:
      type: object
      required: [id, versionId, agentId, name, description, version, sourceType, content, contentHash, enabled, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: Stable Skill package identifier used by Agent binding routes
        versionId:
          type: string
          description: Immutable selected Skill version identifier for this Agent
        agentId:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: string
        sourceType:
          type: string
          enum: [inline, local, git]
        content:
          type: string
        contentHash:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InstallSkillRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 131072
        version:
          type: string
          maxLength: 80
          description: Optional immutable version label; omitted inline installs use a content-derived local label
    McpTool:
      type: object
      required: [name, description, inputSchema, enabled, riskLevel]
      properties:
        name:
          type: string
        description:
          type: string
        inputSchema:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        riskLevel:
          type: string
          enum: [read-only, external-write, destructive]
    McpServer:
      type: object
      required: [id, agentId, name, endpoint, transport, enabled, status, tools, createdAt, updatedAt]
      properties:
        id:
          type: string
        agentId:
          type: string
        name:
          type: string
        endpoint:
          type: string
          format: uri
        transport:
          type: string
          enum: [streamable-http]
        enabled:
          type: boolean
        status:
          type: string
          enum: [unchecked, connected, error]
        protocolVersion:
          type: string
        lastError:
          type: string
        lastCheckedAt:
          type: string
          format: date-time
        tools:
          type: array
          items:
            $ref: "#/components/schemas/McpTool"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateMcpServerRequest:
      type: object
      required: [name, endpoint]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        endpoint:
          type: string
          format: uri
          maxLength: 2048
    Conversation:
      type: object
      required: [id, agentId, title, createdAt, updatedAt]
      properties:
        id:
          type: string
        agentId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, conversationId, role, content, sequence, createdAt]
      properties:
        id:
          type: string
        conversationId:
          type: string
        runId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sequence:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    Run:
      type: object
      required: [id, conversationId, status, createdAt]
      properties:
        id:
          type: string
        conversationId:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        failureCode:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
    RunEvent:
      type: object
      required: [runId, sequence, type, payload, occurredAt]
      properties:
        runId:
          type: string
        sequence:
          type: integer
          format: int64
        type:
          type: string
          enum:
            - run.started
            - tool.started
            - tool.completed
            - tool.failed
            - message.delta
            - message.completed
            - run.failed
            - run.cancelled
        payload:
          type: object
          additionalProperties: true
        occurredAt:
          type: string
          format: date-time
    AgUiRunAgentInput:
      type: object
      required: [threadId, runId, state, messages, tools, context, forwardedProps]
      properties:
        threadId:
          type: string
          description: Existing AegisLink conversation identifier
        runId:
          type: string
          description: Client-generated identifier for this AG-UI run
        parentRunId:
          type: string
        state:
          description: Agent-owned state supplied by the AG-UI client
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AgUiMessage"
        tools:
          type: array
          items:
            type: object
            additionalProperties: true
        context:
          type: array
          items:
            type: object
            additionalProperties: true
        forwardedProps:
          type: object
          additionalProperties: true
    AgUiMessage:
      type: object
      required: [id, role]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [developer, system, assistant, user, tool, reasoning]
        content:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                additionalProperties: true
      additionalProperties: true
    ConversationDetail:
      type: object
      required: [conversation, messages]
      properties:
        conversation:
          $ref: "#/components/schemas/Conversation"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        activeRun:
          $ref: "#/components/schemas/Run"
    BootstrapResponse:
      type: object
      required: [user, agent, model]
      properties:
        user:
          $ref: "#/components/schemas/User"
        agent:
          $ref: "#/components/schemas/Agent"
        model:
          $ref: "#/components/schemas/ModelInfo"
    ModelInfo:
      type: object
      required: [id, driver, name, capabilities]
      properties:
        id:
          type: string
          description: Stable identifier for the configured model profile
        driver:
          type: string
          description: Runtime provider used to create the model
        name:
          type: string
          description: Provider-specific model name
        capabilities:
          type: array
          items:
            type: string
            enum: [streaming, tool-calling]